split main and util
